backend/requirements.txt
django
djangorestframework
mysqlclient
djangorestframework-simplejwt
django-filter
django-celery-beat
django-cors-headers
django-timezone-field
drf-spectacular[sidecar]
django-simpleui
celery==5.4.0
redis
requests
certifi
charset-normalizer
requests_toolbelt
jsonpath
faker
pymysql
pymssql
oracledb
gunicorn
supervisor
gevent
flower
django-celery-results
eventlet